LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-setLocalDescription-offer_interop-2026-expected.txt


FAIL setLocalDescription() with offer not created by own createOffer() should reject with InvalidModificationError assert_equals: expected "InvalidModificationError" but got "OperationError"
